Wrong key for reference field in Commerce admin UI
The foreach() loop that searches node types for reference fields currently looks for the following:

`if (isset($properties['settings']['referenceable_types'][$product_node_type])) { `

However the keys in `$properties['settings']['referenceable_types']` are that of product types, not node types. The result is that this array never returns any field values, because display node types do not exist in this array. Simply looking for the `referenceable_types` array should suffice here, as that indicates the field is an entityreference field.